Refactors PhaseChangeLevel plugin into its own module
